Add support for slave mounts.
Previously, the entire set of mounts within the container were
marked 'private': this means any changes in the host OS would not
show up in the container.

However, in places that use autofs (e.g., distributing software),
we want the autofs daemon (which lives in the host OS) to be able
to make mounts that show up in the container.  To do that, we must
mark the mounts as 'slave': changes in the host show up in the
container, but not vice-versa.

Unfortunately, it seems that it's all-or-nothing: we can't do slave
mounts _only_ for a single bind mount.  Accordingly, I set this to
be a global, container-wide option as opposed to only setting it for
a single bind mount at-a-time.  The default is to use private mounts
(the pre-existing behavior).