Introduce the concept of private scratch directories.
Allow the sysadmin to configure private scratch directories for the
container.  These directories are writable and owned by the user,
but unique to the container instance.

This is handy for providing private /tmp and /var/tmp for jobs.

NOTE: These are applied *after* 'bind path', which might not be the
correct semantics.  Can't see one being obviously-better over the
other.