Add user namespace support / unprivileged mode.
This commit adds support for unprivileged operation _if_ we are using
a chroot instead of a raw disk image (use of loopback devices are still
banned for raw disk images).  Remapping is done so the invoking user
'looks right' in the namespace, but all other file ownership looks strange.

Should work on modern Ubuntu and Fedora variants.  RHEL7 doesn't have
enough backported patches to make this function.  As it's tricky to
determine whether a given host has support for this mode, it must be
explicitly enabled at compile time.