Allow sysadmins to disable user scratch directories.
The sysadmin needs to be able to disable user scratch directories
in case there's a local issue with the feature (it's believed to
always be safe).

NOTE: if scratch directories are disabled and the user requests
them anyway, this is an immediate failure.  I figure that's better
than a silent degradation of the feature.