Catch weird RPM dbpath values
On Debian-family OSes, the rpm command places your rpmdb in $HOME/.rpmdb.
This catches that situation at bootstrap-time, aborts, and advises the
user on how to fix the problem.

NOTE: Since the fix involves changing a file outside the container,
let the user make the change.

NOTE: rpm does have the --rpmdb option, but yum doesn't, so we need to
change a config file.

See singularityware/singularity#241