Allow hyphen in TERM environment variable
Running `ls /lib/terminfo/* | grep \-` on a stock Ubuntu 16.10 Desktop
system shows 22 terminal types that use the hyphen character.  When using
the Terminal in the Unity desktop, the terminal type is set to
`xterm-256color`.

A manual scan of the list shows that the hyphen is the only non-alphanum
character in the list of terminal types.

Closes #577.