always use 'try' for overlayfs functionality.
Finding the module on build host is neither a guarantee it will
exist in the runtime environment, nor is absence of the module
a guarantee overlayfs is not supported (may be compiled into
kernel).
So don't deduce runtime environment from build host,
but always use the new 'try' feature.
For testing, activate that if the module is found.

Signed-off-by: Oliver Freyermuth <o.freyermuth@googlemail.com>