update_token: make scope,service params optional
Updated the `Www-Authenticate` challenge parsing code in the
`update_token` function in the python docker client to make `service`
and `scope` params optional. These params are then conditionally used
when building the token URL based on whether they were present in the
auth challenge. The realm parameter is still required as that's the
endpoint used for auth.

Not all private docker registries provide scope and service params in
the auth challenge. For example, NVIDIA's private registry, nvcr.io,
returns only the bearer realm as part of their proxy authentication flow
and this is supported by the native docker client.

https://github.com/singularityware/singularity/issues/1159#issuecomment-347593106