strip auth header on redirect to different domain
This fix prevents potentially leaking credentials via http redirects to
hosts other than in the original request. This change mimics both curl
and the python "requests" lib's approach to solving this issue. See:

https://github.com/curl/curl/blob/8123560d4496992baecc0c7a05e56ce48e6b6273/lib/http.c#L1666
https://github.com/requests/requests/blob/f60324a3de41c10dd7af17cadd4322c895068a25/requests/sessions.py#L240

At the same time this serves as a fix for docker registries using a
secured S3 backend where the registry redirects to a secure S3 URL.
Without stripping the auth header S3 denies the request with:

```
<?xml version="1.0" encoding="UTF-8"?>
<Error>
  <Code>InvalidArgument</Code>
  <Message>Only one auth mechanism allowed; only the X-Amz-Algorithm
  query parameter, Signature query string parameter or the Authorization
  header should be specified</Message>
  <ArgumentName>Authorization</ArgumentName>
  <ArgumentValue>Bearer REDACTED</ArgumentValue>
  <RequestId>REDACTED</RequestId>
  <HostId>REDACTED</HostId>
</Error>
```

See the following for issues/PRs for more details:
- singularityware/singularity#1159
- singularityware/singularity#1184