Add util/binary to inspect binaries file (ELF)
This patch add a function to get the architecture (x86_64, i386,
x32...) of a specific ELF binary. This can be used to sort binaries
to embed in the container to different files.

This introduces a dependency on libelf for ELF parsing.

It only supports x86 and derivatives (x86-64 and x32) but should
be easily supports more archs if necessary.