add multiarch support to CONTAINLIBS
The previous behavior was to iterate the given libraries, and if
the same library name (but not the same source path) was supplied
twice or more, only the first was bound.

With this patch libraries will be sorted by architecture in sub
directories of /.singularity/libs/. The LD_LIBRARY_PATH is updated
with the used architectures.

If the architecture of a file cannot be determined (unkown arch or
not an ELF file), the file is cpoied directly in libs. If the same
destination should be used twice or more the behavior is the same
as before, only the first is used.

This patch only supports x86_64, i686 and x32. Support for other
arch can be added later.

With this patch using --nv on a host with the 32 and 64 bits
NVIDIA libraries installed will permit to use 32 and 64 bits
application inside the container.