Add compile.sh and build a proto singularity
This commit brings in a small compilation driver that ties in the
compilation of necessary C core library (libsycore.a) and the build of
small go prototype program using the lib.

The go target will change greately but the compile.sh driver will most
likely remain to kick off the project builds.

Signed-off-by: Yannick Cote <yhcote@gmail.com>