edit Singularity license info so that GitHub recognizes it
GitHub uses a library called Licensee to identify a project's license
type. It shows this information in the status bar and via the API if it
can unambiguously identify the license.

This commit makes a couple changes so that Licensee recognizes
Singularity's license type as BSD 3-Clause. It removes the COPYRIGHT and
LICENSE-LBNL.md files so that Licensee correctly identifies LICENSE as
the sole license file. The contents of those files are now stored in a
new "Licensing" section in the README (namely, the notice regarding the
U.S. DOE and the list of copyright holders).

LICENSE itself has been updated so that its contents match the standard-
formatted text of the BSD 3-Clause.

Signed-off-by: Andrea Kao <eirinikos@gmail.com>