Specify bytes-per-inode when creating image
When using image.create to make an overlay image, this will allow
the specifying of the size of an inode.

mke2fs default is 16K per inode, which if you have a lot of small
files, means you would hit the inode limit before you would hit
the space limit.

This PR adds a `-i` option to specify the size in bytes of the
inodes to mkfs.ext3.

For example:
```
$ singularity image.create -i 8192 /tmp/overlay.img
```

Issue: #1759