Directory creation without OverlayFS
Try creating directories with `--contain` even if OverlayFS isn't
supported. The reason for this is because /dev at this point is
a tmpfs, and we can write to it.

So say, for example, we want /dev/infiniband on a system that
doesn't support OverlayFS, we can do:

  `singularity exec -c -B /dev/infiniband [image] [a b c]`

And it will work.