Introduce proof of concept code for packaging plugin objects as SIF files
This commit adds rudimentary functionality for converting the .so file built
into a .sif file. Eventually the plugin .sif will contain its manifest as
an object, as well as the .so file as a raw data partition. Currently, only
the .so file is added but a future commit will also add the manifest.