Add support for platform-specific default values
action_flags.go contains a bunch of flags shared among several commands.
With the new VM support, using a VM should be enabled by default on
darwin and disabled by default on Linux. Move that difference to
build-time by introducing a new function initPlatformDefaults which
should be used to set any platform-specific action flags.

The list of action flags is also platform-specific, so move the list of
supported flags (not the flag definitions) to a target-specific file,
too.

Use cobra's built-in method to detect whether the user passed a flag or
not in order to implement dependent flags (--vm-cpu implicitly enables
--vm). This avoids hard-coding the default values in multiple locations.

TODO: further clean-up of the actionFlags definition?

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>