Print usage information if an invalid flag is passed to singularity
In cobra, rootCmd.Execute() processes the command line and executes the
appropriate command or subcommand. If an invalid flag is passed to the
subcommand, the usage information for the subcommand is printed. If an
invalid subcommand or flag is passed to the root command, the usage
information for the root command is printed. If in invalid subcommand is
passed along with an invalid flag (`singularity -f xxx shell`), cobra
prints an error. Since we are suppressing cobra errors (SilenceErrors),
nothing is printed and singularity returns silently.

Fix this by passing SilenceUsage as well, and take care of printing the
usage string ourselves for all errors.

- Fixes: #2534

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>