Change sylog.Fatalf to sylog.Errorf in plugin enable / disable
sylog.Fatalf will terminate the program. There are (exceptional) valid
use cases for that, but I think at this level, it's better to get into
the habit of using sylog.Errorf, as sylog.Fatalf will call os.Exit, and
os.Exit prevents defer'ed functions from running, which in this case
it probably does not make a difference, but the keyword here is
"probably" (as we don't know, in general, what the calling code is
doing).

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>