Go back to sylog.Fatalf in plugin enable / disable
sylog.Fatalf calls os.Exit, meaning it immediately terminates the
program and does not allow defer'ed functions to run, etc. Since in
general we don't know what the calling code is doing, this is the wrong
thing to do, but it is the existing pattern. So for consistency's sake,
go back to that until we fix this properly everywhere.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>