Fix loading of plugins
Use the meta file to locate the actual plugin binary and load it from
there. Right now the code is trying to load the meta file as an ELF
file.

In order to make sure the plugin can be loaded, include go build tags in
generated config.go file and pass that value to the Go compiler when
building plugins. In this way we guarantee that the plugin matches the
singularity build.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>