Move Go version checks completely to build time
The current method of checking the Go version is implemented partly at
build time and partly at runtime. At build-time the compiler selects
between to versions of the same function, one that produces an error if
the Go version is not supported and one that does not produce an error.
At runtime, this function is called, and the error is checked. This
means that the user has to build, install and run the software in order
to figure out that the Go version is not supported.

With this change the check happens not only at build-time, but it
happens very quickly. Near the end of the output, there's a line that
reads:

    ../internal/pkg/util/goversion/version_check.go:19:9: undefined: __BUILD_REQUIRES_GO_VERSION_1_10_OR_LATER__

which should be clear enough for anyone trying to build singularity on
their own. It has the additional side-effect of causing CI to fail fast,
too.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>