Introduce support for remote add/remove --global flag
We'd like to support the usage of a global remote endpoint config
file which gives users a set of pre-configured SCS endpoints to select
from. The first step taken here is allowing the root user to add and
remove endpoints from a global configuration file. This file is located
at $(buildcfg.SYSCONFDIR)/singularity/remote.yaml.

Updates: #2894

Signed-off-by: Michael Bauer <michael@bauer.dev>