Add remote.Config method SyncFrom to merge two remote Configs
In order to better manage two Config files, we add a SyncFrom
method to remote.Config, as well as a System field to the
remote.EndPoint type. SyncFrom will return a name-collision error
if there is an EndPoint name which exists in both Configs and
the EndPoint in the user-specific Config is not marked as a
System EndPoint (via EndPoint.System). SyncFrom will also update
the URI stored in the user-specific Config if it has changed.

Updates: #2894

Signed-off-by: Michael Bauer <michael@bauer.dev>