Generate 9p Mount Tag on Bind Mount
Generate a random 6 alphabetic char 9p mount tag for all mounts that
are specified with the `-B` option. This means that a user can set
both `-B /tmp -B /var/tmp` and it will work as expected.

Previously `tmp` would have been used as the mount tag for both
of the bind mounts.