Remove early debug output in internal/pkg/plugin
Since the init functions for internal/pkg/plugin are running before
those for sylog, the default log level is still at "debug" and these
messages show up whenever a plugin is enabled (irrespective of the
actual log level set on the command line).

Since these are not critical, remove them. If they are deemed to be
useful later, they can be restored after the underlying issue is fixed.

Fixes: #3133

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>