Update Bind Mounts
Modify how bind mounts are getting passed. We were hard limiting the
number to 8, and that is now at 48 from the Mac host (assuming my
math is right... which I don't guarantee)

Start the bind mount slot at 26, and increase the slot number every
time we hit 8 mounts.

Error if we surpass the max slot.