Conditionally pass -comp gzip to mksquashfs
For versions of mksquashfs that support it, we need to pass -comp gzip
(because some versions out there default to a different compression
algorithm). Since some (older?) versions of mksquashfs do not support
the -comp flag, we have to test this at runtime: try to use it and if
that fails, assume that mksquashfs does not support it.

Fixes: #2717

Signed-off-by: Ian <iankane@umich.edu>
Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>