Add initial support for pushing SIF to OCI registries
This work is currently leveraging github.com/deislabs/oras to
help facilitate this. However, it looks like we need to implement
a content.Provider since the oras package doesn't quite fulfill
what we need. In its current state, we either need to write the
config JSON into a file, or read the entirety of the SIF file into
memory to allow oras to push it. We can fix this before we merge
this PR into Singularity.

Signed-off-by: Michael Bauer <michael@bauer.dev>