Add EnsureImage function to make sure a test image exists
Instead of simply creating an image for every run of e2e tests, move the
test image creation to a separate package and provide a function
(EnsureImage) that makes sure the image exists before the test that
needs it starts to run.

Add LoadEnv to avoid copy-pasting the same code to load the environment.

- Fixes #3409

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>