Make sure Go modules are disabled when checking Go version
Since we are running a simple program to check that the Go version
meets the minimum requirements, we should not need any modules listed in
go.mod.

This takes care of a problem in CentOS 6 where, seemingly, the git
version that ships with it does not work well with the Go version that
ships with it _when modules are enabled_ (then the go tool tries to go
to the network to donwload modules).

The rest of the code should be fine as it's using the vendor directory
(by way of passing -mod=vendor to the go tool).

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>