Add syfs.ConfigDir() to obtain singularity's config directory
There are multiple locations in the code that have their own method to
define where the singulary configuration directory exists, all trying to
do more or less the same thing.

Move that code to a single location that can be reused by the rest of
the code. Call the package "syfs" to express that this represents
singularity's file system layout.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>