Create scripts/go-test script during build
scripts/go-test is simply a wrapper for go test that uses the same flags
as the singularity build (-tags, -mod, etc), and also passes flags that
make sense in this context (-count=1, -cover, etc).

Use a Go program to do the replacement in scripts/go-test.in in order to
avoid having to mess around with quotes and whatnot.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>