Allow arbitrary locations when pushing to oras
The current implementation requires the user to do something like this:

	cd /some/path
	singularity push file.sif oras://localhost:5000/test:latest

and does not support this:

	singularity push /some/path/file.sif oras://localhost:5000/test:latest

in other words, the local file must be in the current working directory.

Fix the call to:

	store.Add(name, mediaType, path)

so that the first argument is simply a filename (without a directory),
while the third argument is the path specified by the user.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>