Do not manipulate "environ" directly, use C API
Reading thru the docs, it seems the environ variable is not meant to be
used to _remove_ or _change_ values in the environment, but only to
iterate over them.

A recent change introduced the "cleanenv" function, which deletes all
environment variables except SINGULARITY_MESSAGELEVEL. It does this by
setting the pointers in the environ array to NULL. While this seems to
work with some implementations of the C library, with others it's
causing a segmentation fault during the initialization of the C library.

To avoid assuming a particular behavir, change the cleanenv function so
that it obtains the current value of SINGULARITY_MESSAGELEVEL, clears
all the environment variables using the clearenv function and sets
SINGULARITY_MESSAGELEVEL back to its previous value using setenv.

Since the original change did not specify the conditions under which a
bug was observed, this has received only limited testing. All the
existing tests are passing with this change.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>