Clear the cache by deleting the individual caches
Since it's possible for users to nuke their home directory (or in
general other directories in the system that they have access to) by
setting the environment variable SINGULARITY_CACHEDIR to the (in)correct
value, reimplement the way the "cache clean --all" command works by
iterating over the known cache types. This creates a maintainance burden
since the list of known types has to be kept in sync in multiple places,
but this chage is trying to be as small as possible while still being
correct.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>