Add support for numbered variables
To use multipe instances of the same variable, introduce an indexing
system (sort of like an array).
In the validHeaders table, the number index is represented by an
'&n' at the end of the var name:
ie:
    "foo&n": true,
this will match variables of the form:
    foo0: somestring
in the definition file. The value can be referenced from the Header
map using "foo<N>" as key (<N> replaced by the matching integer).

Signed-off-by: Egbert Eich <eich@suse.com>