Rework cache listing and cleaning code
Remove some of the code repetition by moving all the common code to a
single location.

Replace --summary with --verbose, so that by default only a summary is
printed. --verbose adds a listing of the individual entries.

Add --force to cache clean, which is required to actually clean the
entries.

Remove the "blobSum" type, as it's used to request a summary for the
blobs, which is the same thing --summary does when passing the -T blob
flag.

Make the default -T value for list and clean the same, namely "all".

Fixes: #3718, #3717

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>