Add mandatory argument to RunSingularity to specify binary to use
RunSingularity is using "singularity" as the command it wants to run,
which means it's picking up singularity from the user's PATH, which
might or might not correspond to the singularity binary that matches the
code to be tested (e.g. mconfig --prefix=/usr/local/test, and then run
run the E2E suite without adding /usr/local/test/bin to the path).

Fix this by adding a cmdPath argument to RunSingularity which specifies
where the singularity binary is located. This plays nice with the
existing code that records the location of this binary in the
E2E_CMD_PATH environment variable. It's the responsability of the test
calling RunSingularity to pass the correct value derived from this
environment variable.

Add a helper method to TestEnv called RunSingularity that does *not*
take the cmdPath argument. Since TestEnv already has a CmdPath field,
use that to call the standalone RunSingularity function. This obscures
what's going on, but reduces the chances of calling the wrong binary.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>