Pass test environment information as function arguments
Originally the intention was to run each test as a separate program
(there's some important amount of rewriting required to do that).
Because of this some test parameters are being passed in the
environment, in a way that is not immediately obvious. Since that
pattern is causing some confusion and people are outright ignoring it in
some cases, remove it and replace it by adding a test context that
contains all the relevant test information (path to singularity, test
directory, test image, etc).

In order to keep the diff to a minimum, this change has been made almost
mechanically, so there's room for improvement in the future.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>