Disable tests or test cases that access singularity hub
While singulariy hub is down, disable the tests that try to access it.
Every disabled test or test case has been marked with a comment
indicating the reason for disabling it is the fact that singularity hub
is down. This is done in order to avoid confusion in case the code is
reworked before the tests can be reenabled again.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>