Move call to e2e.EnsureImage to the tests that need it (#3872)
EnsureImage creates the test image if it's missing. This is done to
avoid creating the image multiple times for each test.

The call to EnsureImage MUST be in the tests that require said image in
order to avoid paying the cost of creating the image when it is not
required: many tests do not require this image and can operate without
it, and while developing a test or working on individual failure cases
it's useful to be able to run a single test repeatedly. If said test
requires the image, there's no way around it, but if the test does not
require the test image it does not make sense to build it over and over
again.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>
Co-Authored-By: Eduardo Arango <eduardo@sylabs.io>