Skip e2e tests for pull requests based on their changes
We only want to run e2e tests under the following condtions:

- Changes are made to the `e2e` directory
- Changes are made to the `vendor` directory
- Merging a pull request into `master`

This requires the use of the `$CIRCLE_COMPARE_URL` environment
variable provided by the orb `iynere/compare-url`.

https://circleci.com/orbs/registry/orb/iynere/compare-url

In earlier versions of circleci, that environment variable was
already present without the need for the orb. Consider this a
short-term fix.

- Fixes #3430
- Fixes #3806