Check for cryptsetup at build time
Currently the location of cryptsetup is hardcoded to /sbin/cryptsetup.
Check for the actual path at build time, and record the result in
buildcfg. At runtime, the path is retrieved from the configuration file
first, and if that's empty, it falls back to the path recorded at build
time. This gives the system administrator full control over which
cryptsetup program is used.

Fixes: #3819

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>