Rename GetCryptDevice to Open
GetCryptDevice is looking for a device name that can be used during the
opening of the crypt device, but it's doing so by trying to open the
block device and assign it a name, effectively opening it.

Instead of modifying the function to do something different, rename it
to reflect what it's doing.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>