Rework copyDeviceContents to deal with short reads and writes
It's possible for a read or write operation to get interrupted and not
read or write the full buffer. Also it's possible that there isn't
enough data available and this might result in a short read.

Rework copyDeviceContents to deal with those situations.

Also use a slightly larger buffer because it's faster.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>