Support opening images using a private PEM key
Add support for opening (run, exec, etc) images that contain an
encrypted decryption key. The decryption key is encrypted using a public
key in PEM format and it's decrypted using a private key in PEM format.

For now, assume that the first Generic partition linking to the system
partition contains the encrypted decryption key.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>