Fix buildLocalImage so that it does not step over itself
buildLocalImage is trying to test that building images from several
sources works as intended. It does so by starting with one image and
turning it into a different kind of image. The code is such that it
"knows" when the image it's trying to build is already there and
therefore passes the "--force" flag to avoid an error... except when it
doesn't and the image is already there, which can happen if more than
one of the subtests happen to run at the same time or in a different
order than the test expects them to.

Fix this by removing all the complexity and test what the test is trying
to test in the first place: building a single image out of a
specification file (source, sandbox or definition file). Each subtest
uses either the test image that is already available to all the tests or
a sandbox that is derived from the test image.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>