Introduce SingularityCmdPath to catch migration errors
When the signature of RunSingularity changed from:

	func(*testing.T, string, ...SingularityCmdOps)

to:

	func(*testing.T, string, ...SingularityCmdOps)

the problem is that it didn't change. The only thing that changed was
the meaning of the second argument.

Change the type of the second argument SingularityCmdPath so that the
compiler produces an error with the old code and make it easier to catch
the mistake.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>