Add stack traces to test errors
In order to make it easier to understand what's failing, use
github.com/pkg/errors to record the closest location to where an error
was reported and use that when calling t.Fatal or equivalent functions.
In this way a stack is captured at the location where the error is
receving and it's printed out as part of the error message.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>