Remove the dependency on homebrew to install go
homebrew runs an update before installing any packages, which can
take awhile. Since we are only building and linting on macos (and
not doing any functional testing), remove homebrew as a dependency
to reduce build times.

In the future, we may add homebrew back when more dependencies
besides go are required.

Fixes: #4106