Add "Profile" as steppingstone for testing --fakeroot
As part of testing the --fakeroot functionality, sometimes it's
necessary to run the same singularity command under multiple
"personalities" or "profiles". For example "singularity build" can be
executed with privilege or as "singularity build --fakeroot", and they
should both succeed. It's also expected that the command fails if run as
a regular user.

As part of end-to-end testing, all these personalities need to be
tested, and the "Profile" concept introduced here helps with that. The
RunSingularity driver needs to be extended to support this.

Most of this comes from Cedric; this is mostly a reorganization of that
code, with some documentation sprinkled into it.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>