Add go-generate script
Add a scripts/go-generate script in the same vein as go-test to make
sure that we use the same build flags as when building singularity
itself.

The script can later be used to generate the Go code _before_ building
singularity, which allows us to run the code linter before building
singularity (otherwise the linter fails because of missing files).

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>