Take extra build tags in go-test
go-test passes its arguments verbatim to `go test`. If you pass `-tags`
it fails because there's already a `-tags` flag in the script and `go
test` does not accept multiple `-tags` flags (this would have been
nice).

So pay attention to the passed flags and build the value of `-tags` from
the ones that are passed to `go-test` and the ones the script has built
in (from mconfig).

This allows the user to pass additional flags to `go test` if necessary.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>