Add `integration_test` build tag to integration tests
Instead of keeping an implicit record of which tests are integration
tests are which are something else, make it explicit using the existing
Go mechanisms.

Specifically add an `integration_test` build tag to the integration
tests and pass `-tags integration_tests` to the test runs that wish to
include these tests in their runs.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>