Don't prune integration and E2E tests by hand anymore
The purpose of this:

    `$(GO) list ./... | grep -v "cmd/singularity\|pkg/network\|e2e"`

is to manually prune the directories that contain integration and E2E
tests from the unit test run. While this works in the current state of
the repository, note that this is a brittle mechanism, as it uses loose
regular expressions to catch the undesired directories.

Now that the integration and E2E are properly tagged using the
`integration_test` and `e2e_test` build tags, it's possible to remove
this pruning and simply request everything to be tested. By not passing
the corresponding flags, the undesired tests will be automatically
pruned \o/

This has the extra benefit that it will become obvious which code is
lacking tests.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>