Move internal/pkg/util/exec/pipe* to internal/pkg/util/starter and code to execute starter was refactored to remove some code redundancies disseminate in CLI and internal packages.
Also fix issue during image cleanup when using fakeroot with
SUID workflow, because CleanupContainer method is called by
master process when fakeroot+SUID workflow are in use this
method is executed outside of the fakeroot context which means
that under some circumstances the image removal may fail with
a permission denied error. To fix the issue the fakeroot engine
is called in order to execute /bin/rm command.

Co-Authored-By: Sasha Yakovtseva <sashayakovtseva@gmail.com>