Mark some tests as "not short"
Add `if testing.Short()` to the slower tests (above one second) to try
to keep unit test runtime under control. Some of these tests are
exceedingly long (over a minute).

This is _not_ changing anything in terms of the tests that run or don't
run, as the required flag to `go test` (`-short`) is not passed at the
moment.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>