Reformat imgBuildTests.buildFrom
It took some heavy editing to notice that the individual tests weren't
being run as subtests, which has a weird interaction with the e2e.PreRun
where tests are skipped based on the availability of a particular
program in the host.

It also took some heavy editing to notice that the BusyBox test was
missing the --sandbox option, which is required to get the ImageVerify
helper to do its thing.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>