Use tables to drive E2E tests (#4432)
* Use tables to drive E2E tests

A number of E2E suites have this structure:

	return func(t *testing.T) {
		// some thing
		t.Run("testSomething", c.testSomething)
		// another thing
		t.Run("testAnotherThing", c.testAnotherThing)
		// one more
		t.Run("testOneMoreThing", c.testOneMoreThing)
	}

CamelCase is not really necessary here, and it makes things a little
harder to read. Replace those by short sentences that hopefully describe
the group of tests represented by the corresponding function.

The sequence of calls to t.Run is a little repetitive. Replace that by a
map with the test name and the test function and a new helper function
to call the functions.

The result looks like this:

	return testhelper.TestRunner(map[string]func(*testing.T){
		"another thing":  c.testAnotherThing, // another thing
		"one more thing": c.testOneMoreThing, // one more
		"something":      c.testSomething,    // some thing
	})

The tests are sorted alphabetically by their description, so some care
should be paid to NOT create tests that must run in sequence.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>

* Update e2e/internal/testhelper/doc.go

Co-Authored-By: Sasha Yakovtseva <sashayakovtseva@gmail.com>