Get capabilities from /proc/self/status in E2E tests
The current implementation of testSecurityPriv and testSecurityUnpriv
uses ping to infer the status of the cap_net_raw capability, as a proxy
for testing whether --keep-privs, --no-privs and --drop-caps are doing
the right thing.

Change this to directly reading /proc/self/status to obtain the
effective set of capabilities in the container. The downside of this is
that the complete set of capabilities might change, so this test might
break in the future if something is added to the set of known
capabilities. Punt that problem for now.

This removes the need to download an image, making the test faster.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>