Use go-generate script to generate config.go
go-generate was added a while ago, but it's not yet used. Its purpose is
to avoid an error where the go binary has been found by mconfig in an
unusual location, but the "go generate" step doesn't use that binary and
instead tries to locate it again on the PATH.

Since the generated file is not going to change, it should be fine to
pull this change and continue without removing builddir first.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>