Pass go's path to gengodep (#4515)
* Pass go's path to gengodep

When mconfig runs, it determines the absoluate path to the go compiler
to be used. The gengodep script is used in the generated Makefile in
order to generate dependencies. Because of the way this has been
implemented, it runs every time the Makefile is used, even if the
compilation is already done. Since gengodep is not using the path that
mconfig determined, when running "make install" it's possible that the
"go" program is no longer found via PATH.

Fix this by passing the absoluate path to gengodep.

To avoid weird errors, add an argument to gengodep to specify the
expected "version" of the script. If the makefile is updated without
updating the script, or if the script is updated without updating the
Makefile (the more common case), a message will be printed and an error
will be returned. The error is not enough to make the compilation fail,
but it should be visible enough to let the user know what needs to
happen.

Fixes #3547.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>
Co-Authored-By: Jason Stover <jason.stover@gmail.com>