Fix #4524 - set permissions on rootless sandboxes
This fix addresses #4524, where the recent change to use the umoci
extracter for OCI layers can lead to sandboxes that cannot be mv'd across
filesystems (leading to a build failure), or deleted/modified by the
user.

Return to the <=3.3 behaviour here, by modifying permissions on the
temporary rootfs sandbox once it is extracted, and before the mv to
its final location.